Lift MAX_ASTEROIDS from 255 to 2048 with uint16 wire protocol
Wire protocol v2: asteroid indices are now uint16 (2 bytes) instead
of uint8. Count field also uint16. Record size: 31 bytes (was 30).
Old clients cannot connect — controlled deploy.

- MAX_ASTEROIDS = 2048 (was 255)
- ASTEROID_RECORD_SIZE = 31: [index:u16][flags:u8][7×f32]
- Message header: [type:u8][count:u16] (3 bytes, was 2)
- Client deserializer reads uint16 count + index
- Full snapshot only sends active slots (skip empty at 2048)
- Interpolation arrays grow to 2048 (320KB, in global game_t)
- Tests updated for new wire format offsets

4 spread players can now have ~500 unique asteroids each without
pool starvation. Chunk terrain + per-player relevance + 2048 pool
= infinite discoverable asteroid fields.

+++ src/net.h
@@ -50,9 +50,9 @@ typedef struct {
     bool active;
 } NetPlayerState;
 
-/* Packed asteroid state for world sync (30 bytes per asteroid). */
+/* Packed asteroid state for world sync (31 bytes per asteroid). */
 typedef struct {
-    uint8_t index;      /* asteroid slot 0-254 */
+    uint16_t index;     /* asteroid slot 0-2047 */
     uint8_t flags;      /* bit0=active, bit1=fracture_child, bits2-4=tier(3), bits5-7=commodity(3) */
     float x, y;         /* position */
     float vx, vy;       /* velocity */
